+++ src/Exceptions/AbilityProviderNotFoundException.php
@@ -1,0 +1,20 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Bearer\Exceptions;
+
+use function sprintf;
+
+final class AbilityProviderNotFoundException extends AbstractAbilityProviderNotRegisteredException
+{
+    public static function forName(string $name): self
+    {
+        return new self(sprintf('Ability provider "%s" is not registered.', $name));
+    }
+}