+++ src/Exceptions/AbstractAbilityProviderNotRegisteredException.php
@@ -1,0 +1,17 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Bearer\Exceptions;
+
+use InvalidArgumentException;
+
+abstract class AbstractAbilityProviderNotRegisteredException extends InvalidArgumentException implements BearerExceptionInterface
+{
+    // Abstract base class - use concrete implementations.
+}